requirements.txt
fastapi==0.78.0
tensorflow==2.9.1
uvicorn==0.18.2
pyyaml==6.0
datasets==2.3.2
